Add unit tests for movies view model.
